Fixed cast warnings in c++
git-svn-id: svn://136.177.114.72/svn_GW/phreeqc/trunk@1008 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/basicsubs.c
@@ -787,7 +787,7 @@ int match_elts_in_species (char *name, char *mytemplate)
 	 */
 	case_no = 0; 
 	if (template1[0] == '*') case_no = 1;
-	l = strlen(template1);
+	l = (int) strlen(template1);
 	if (template1[l - 1] == '*') {
 		if (case_no != 1) {
 			case_no = 2;
